Rewrite register method of Bus
I changed the implementation of the bus a bit because there was an
unchecked type cast. However, this is because problems with Type Erasure
in Java. It is impossible to rewrite the code in a different way that
does not show a warning.
In reality the type cast will always be possible.
I inserted a 'SuppressWarnings' annotation in the code to stop the
compiler from reporting it.