Add PluginCommunicator test and implementation
I added a test and an implementation for one public method in the plugin
communicator.
I also had to update gradle to a newer version.
Furthermore, for now I added a "BasicProcessedText" class that extends
the abstract base class "ProcessedText". Possibly this can be used for
the Basic plugin that is going to be built.