Integrate kernel into plugin opening
I rewired opening a plugin in a way that it effectively uses the kernel.
This commit is UNTESTED! I could not verify if it worked because I do
not know the steps of generating the different artifacts and apks.
As a result, do not use this commit to revert because it is unstable.