Implement retrieve cache methods and tests
The methods to retrieve a file from the cache are implemented, along
with tests. However, because the method to clear the cache is not yet
implemented, the tests will probably not always work.

Also, a method in the PluginObject has been changed to work more
generically without having to override it in the subclasses.