Make Kernel a singleton class
Initially it was decided to not use singletons in the code.
The main reason for this was because it introduces global state, which
is bad for testability. However, because there only needs to be one unique
kernel instance in the entire application, that needs to be accessible
from different places, the Kernel class has been made singleton to make
this possible.

This decision has been communicated with and approved by the test
manager @lrenders.